backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
supabase
Pillow
qrcode[pil]
python-dotenv
httpx
anthropic
google-cloud-vision
requests
